refactor: rename NPM utils to PackageManager
